+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/tests/fixtures/checkbox.ts
