--- app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/tests/fixtures/defaultSubmission.ts
@@ -1,208 +1,0 @@
-import { Submission } from "@lib/types";
-
-export const defaultSubmission = <Submission>{
-  form: {
-    id: "cm3yrwzum0000pznaguteuemv",
-    updatedAt: "Wed Nov 27 2024 11:14:04 GMT-0500 (Eastern Standard Time)",
-    closedDetails: undefined,
-    form: {
-      groups: {
-        end: {
-          name: "End",
-          titleEn: "Confirmation page",
-          titleFr: "Confirmation page",
-          elements: [],
-        },
-        start: {
-          name: "Start",
-          titleEn: "Start page",
-          titleFr: "Start page",
-          elements: ["1", "2", "5", "3", "4"],
-          nextAction: "review",
-        },
-        review: {
-          name: "Review",
-          titleEn: "End (Review page and Confirmation)",
-          titleFr: "End (Review page and Confirmation)",
-          elements: [],
-          nextAction: "end",
-        },
-      },
-      layout: [1, 2, 5, 3, 4],
-      titleEn: "Form with tags",
-      titleFr: "",
-      elements: [
-        {
-          id: 1,
-          tag: "tag1",
-          type: "textField",
-          properties: {
-            tag: "tag1",
-            choices: [{ en: "", fr: "" }],
-            titleEn: "A question",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 2,
-          tag: "tag2",
-          type: "textField",
-          properties: {
-            tag: "tag2",
-            choices: [{ en: "", fr: "" }],
-            titleEn: "Another question",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 5,
-          type: "checkbox",
-          properties: {
-            tag: "e379a289-b5a5-480b-818a-c7070fd900c8",
-            choices: [
-              { en: "one", fr: "" },
-              { en: "two", fr: "" },
-              { en: "three", fr: "" },
-            ],
-            titleEn: "Checkboxes",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 3,
-          type: "formattedDate",
-          properties: {
-            tag: "date-tag",
-            choices: [{ en: "", fr: "" }],
-            titleEn: "a date",
-            titleFr: "",
-            dateFormat: "DD-MM-YYYY",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 4,
-          type: "dynamicRow",
-          properties: {
-            tag: "961ec6b4-af3a-4739-a332-0d70a5de50e4",
-            choices: [{ en: "", fr: "" }],
-            titleEn: "repeating set",
-            titleFr: "",
-            dynamicRow: {
-              rowTitleEn: "Instance",
-              rowTitleFr: "Instance",
-              addButtonTextEn: "Add",
-              addButtonTextFr: "Ajouter",
-              removeButtonTextEn: "Remove",
-              removeButtonTextFr: "Supprimer",
-            },
-            validation: { required: false },
-            subElements: [
-              {
-                id: 401,
-                type: "textField",
-                properties: {
-                  tag: "961ec6b4-af3a-4739-a332-0d70a5de50e4",
-                  choices: [{ en: "", fr: "" }],
-                  titleEn: "A question",
-                  titleFr: "",
-                  validation: { required: false },
-                  subElements: [],
-                  descriptionEn: "",
-                  descriptionFr: "",
-                  placeholderEn: "",
-                  placeholderFr: "",
-                },
-              },
-              {
-                id: 402,
-                type: "textField",
-                properties: {
-                  tag: "961ec6b4-af3a-4739-a332-0d70a5de50e4",
-                  choices: [{ en: "", fr: "" }],
-                  titleEn: "another",
-                  titleFr: "",
-                  validation: { required: false },
-                  subElements: [],
-                  descriptionEn: "",
-                  descriptionFr: "",
-                  placeholderEn: "",
-                  placeholderFr: "",
-                },
-              },
-              {
-                id: 403,
-                type: "checkbox",
-                properties: {
-                  tag: "4c4e39f8-4fd8-44fa-9cbd-359e5768eb1d",
-                  choices: [
-                    { en: "one", fr: "" },
-                    { en: "two", fr: "" },
-                    { en: "three", fr: "" },
-                  ],
-                  titleEn: "Check some boxes",
-                  titleFr: "",
-                  validation: { required: false },
-                  subElements: [],
-                  descriptionEn: "",
-                  descriptionFr: "",
-                  placeholderEn: "",
-                  placeholderFr: "",
-                },
-              },
-            ],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-      ],
-      confirmation: {
-        descriptionEn: "",
-        descriptionFr: "",
-        referrerUrlEn: "",
-        referrerUrlFr: "",
-      },
-      groupsLayout: [],
-      introduction: { descriptionEn: "", descriptionFr: "" },
-      privacyPolicy: { descriptionEn: "", descriptionFr: "" },
-    },
-    isPublished: false,
-    securityAttribute: "Protected A",
-  },
-  responses: {
-    "1": "test",
-    "2": "test",
-    "3": '{"DD":1,"MM":1,"YYYY":1900}',
-    "5": '{"value":["one","two"]}',
-    "4-0-0": "asdf",
-    "4-0-1": "asdf",
-    "4-0-2": '{"value":["two","three"]}',
-    formID: "cm3yrwzum0000pznaguteuemv",
-    securityAttribute: "Protected A",
-  },
-};
