--- app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/tests/fixtures/defaultSubmissionResult.ts
@@ -1,7 +1,0 @@
-export const defaultSubmissionResult = {
-  "1": "test",
-  "2": "test",
-  "3": '{"DD":1,"MM":1,"YYYY":1900}',
-  "4": [{ "0": "asdf", "1": "asdf", "2": ["two", "three"] }],
-  "5": ["one", "two"],
-};
