--- app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/tests/fixtures/kitchenSinkSubmission.ts
@@ -1,351 +1,0 @@
-import { Submission } from "@lib/types";
-
-export const kitchenSinkSubmission = <Submission>{
-  form: {
-    id: "cm40dwxmt0002pznalkm7dl9b",
-    updatedAt: "Wed Nov 27 2024 16:23:36 GMT-0500 (Eastern Standard Time)",
-    closedDetails: undefined,
-    form: {
-      groups: {
-        end: {
-          name: "End",
-          titleEn: "Confirmation page",
-          titleFr: "Confirmation page",
-          elements: [],
-        },
-        start: {
-          name: "Start",
-          titleEn: "Start page",
-          titleFr: "Start page",
-          elements: ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12"],
-          nextAction: "review",
-        },
-        review: {
-          name: "Review",
-          titleEn: "End (Review page and Confirmation)",
-          titleFr: "End (Review page and Confirmation)",
-          elements: [],
-          nextAction: "end",
-        },
-      },
-      layout: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
-      titleEn: "Kitchen Sink",
-      titleFr: "",
-      elements: [
-        {
-          id: 1,
-          type: "textField",
-          properties: {
-            choices: [{ en: "", fr: "" }],
-            titleEn: "Short answer",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 2,
-          type: "textArea",
-          properties: {
-            choices: [{ en: "", fr: "" }],
-            titleEn: "Long answer",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 3,
-          type: "radio",
-          properties: {
-            choices: [
-              { en: "one", fr: "" },
-              { en: "two", fr: "" },
-              { en: "three", fr: "" },
-            ],
-            titleEn: "Radios",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 4,
-          type: "checkbox",
-          properties: {
-            choices: [
-              { en: "one", fr: "" },
-              { en: "two", fr: "" },
-              { en: "three", fr: "" },
-            ],
-            titleEn: "Checkboxes",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 5,
-          type: "dropdown",
-          properties: {
-            choices: [
-              { en: "one", fr: "" },
-              { en: "two", fr: "" },
-              { en: "three", fr: "" },
-            ],
-            titleEn: "Dropdown",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 6,
-          type: "combobox",
-          properties: {
-            choices: [
-              { en: "one", fr: "" },
-              { en: "two", fr: "" },
-              { en: "three", fr: "" },
-              { en: "four", fr: "" },
-              { en: "five", fr: "" },
-              { en: "six", fr: "" },
-              { en: "seven", fr: "" },
-              { en: "eight", fr: "" },
-              { en: "nine", fr: "" },
-              { en: "ten", fr: "" },
-            ],
-            titleEn: "Searchable list",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "Start typing to narrow down the list.",
-            descriptionFr: "Commencez à taper pour réduire la liste.",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 7,
-          type: "checkbox",
-          properties: {
-            choices: [
-              { en: "Condition 1", fr: "Condition 1" },
-              { en: "Condition 2", fr: "Condition 2" },
-              { en: "Condition 3", fr: "Condition 3" },
-            ],
-            titleEn: "I agree to:",
-            titleFr: "J'accepte :",
-            validation: { all: true, required: true },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 8,
-          type: "formattedDate",
-          properties: {
-            choices: [{ en: "", fr: "" }],
-            titleEn: "Date",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 9,
-          type: "combobox",
-          properties: {
-            titleEn: "Department or agency",
-            titleFr: "Ministère ou organisme",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "Start typing to narrow down the list",
-            descriptionFr: "Commencez à taper pour réduire la liste",
-            placeholderEn: "",
-            placeholderFr: "",
-            managedChoices: "departments",
-          },
-        },
-        {
-          id: 10,
-          type: "textField",
-          properties: {
-            choices: [{ en: "", fr: "" }],
-            titleEn: "Number",
-            titleFr: "",
-            validation: { type: "number", required: false },
-            subElements: [],
-            descriptionEn: "Enter a number",
-            descriptionFr: "Saisissez un chiffre",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 11,
-          type: "fileInput",
-          properties: {
-            choices: [{ en: "", fr: "" }],
-            titleEn: "File upload",
-            titleFr: "",
-            validation: { required: false },
-            subElements: [],
-            descriptionEn: "Files must not exceed 3.5 MB in total.",
-            descriptionFr: "Les fichiers ne doivent pas dépasser 3.5 Mo au total.",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-        {
-          id: 12,
-          type: "dynamicRow",
-          properties: {
-            choices: [{ en: "", fr: "" }],
-            titleEn: "A repeating set",
-            titleFr: "",
-            dynamicRow: {
-              rowTitleEn: "Instance",
-              rowTitleFr: "Instance",
-              addButtonTextEn: "Add",
-              addButtonTextFr: "Ajouter",
-              removeButtonTextEn: "Remove",
-              removeButtonTextFr: "Supprimer",
-            },
-            validation: { required: false },
-            subElements: [
-              {
-                id: 1201,
-                type: "textField",
-                properties: {
-                  choices: [{ en: "", fr: "" }],
-                  titleEn: "Short answer",
-                  titleFr: "",
-                  validation: { required: false },
-                  subElements: [],
-                  descriptionEn: "",
-                  descriptionFr: "",
-                  placeholderEn: "",
-                  placeholderFr: "",
-                },
-              },
-              {
-                id: 1202,
-                type: "radio",
-                properties: {
-                  choices: [
-                    { en: "one", fr: "" },
-                    { en: "two", fr: "" },
-                    { en: "three", fr: "" },
-                  ],
-                  titleEn: "Radios",
-                  titleFr: "",
-                  validation: { required: false },
-                  subElements: [],
-                  descriptionEn: "",
-                  descriptionFr: "",
-                  placeholderEn: "",
-                  placeholderFr: "",
-                },
-              },
-              {
-                id: 1203,
-                type: "checkbox",
-                properties: {
-                  choices: [
-                    { en: "one", fr: "" },
-                    { en: "two", fr: "" },
-                    { en: "three", fr: "" },
-                  ],
-                  titleEn: "Checkboxes",
-                  titleFr: "",
-                  validation: { required: false },
-                  subElements: [],
-                  descriptionEn: "",
-                  descriptionFr: "",
-                  placeholderEn: "",
-                  placeholderFr: "",
-                },
-              },
-              {
-                id: 1204,
-                type: "formattedDate",
-                properties: {
-                  choices: [{ en: "", fr: "" }],
-                  titleEn: "Date",
-                  titleFr: "",
-                  validation: { required: false },
-                  subElements: [],
-                  descriptionEn: "",
-                  descriptionFr: "",
-                  placeholderEn: "",
-                  placeholderFr: "",
-                },
-              },
-            ],
-            descriptionEn: "",
-            descriptionFr: "",
-            placeholderEn: "",
-            placeholderFr: "",
-          },
-        },
-      ],
-      confirmation: { descriptionEn: "", descriptionFr: "", referrerUrlEn: "", referrerUrlFr: "" },
-      groupsLayout: [],
-      introduction: { descriptionEn: "", descriptionFr: "" },
-      privacyPolicy: { descriptionEn: "", descriptionFr: "" },
-      lastGeneratedElementId: 12,
-    },
-    isPublished: false,
-    securityAttribute: "Protected A",
-  },
-  responses: {
-    "1": "answer",
-    "2": "long one",
-    "3": "one",
-    "4": '{"value":["one","two"]}',
-    "5": "two",
-    "6": "four",
-    "7": '{"value":["Condition 2","Condition 3","Condition 1"]}',
-    "8": '{"YYYY":1900,"MM":1,"DD":1}',
-    "9": "Canada Border Services Agency",
-    "10": "12345",
-    "12-0-0": "short answer",
-    "12-0-1": "two",
-    "12-0-2": '{"value":["one","two"]}',
-    "12-0-3": '{"YYYY":1900,"MM":1,"DD":1}',
-    "12-1-0": "another",
-    "12-1-1": "three",
-    "12-1-2": '{"value":["two"]}',
-    "12-1-3": '{"YYYY":1900,"MM":1,"DD":1}',
-    formID: "cm40dwxmt0002pznalkm7dl9b",
-    securityAttribute: "Protected A",
-  },
-};
