--- app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/tests/fixtures/kitchenSinkSubmissionResult.ts
@@ -1,16 +1,0 @@
-export const kitchenSinkSubmissionResult = {
-  "1": "answer",
-  "2": "long one",
-  "3": "one",
-  "4": ["one", "two"],
-  "5": "two",
-  "6": "four",
-  "7": ["Condition 2", "Condition 3", "Condition 1"],
-  "8": '{"YYYY":1900,"MM":1,"DD":1}',
-  "9": "Canada Border Services Agency",
-  "10": "12345",
-  "12": [
-    { "0": "short answer", "1": "two", "2": ["one", "two"], "3": '{"YYYY":1900,"MM":1,"DD":1}' },
-    { "0": "another", "1": "three", "2": ["two"], "3": '{"YYYY":1900,"MM":1,"DD":1}' },
-  ],
-};
