--- app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/tests/fixtures/pagedFormSubmissionResult.ts
@@ -1,10 +1,0 @@
-export const pagedFormSubmissionResult = {
-  "1": "answer",
-  "2": ["one", "two"],
-  "3": "another",
-  "4": ["two", "three"],
-  "5": [
-    { "0": "Lorem ipsum", "1": ["two"] },
-    { "0": "Hola mi amigo", "1": ["two", "three"] },
-  ],
-};
